+++ src/main/java/gtclassic/recipe/GTRecipeMods.java
@@ -211,6 +211,7 @@ public class GTRecipeMods {
 		/** Forestry Sub Module, last so bees are always last **/
 		if (GTConfig.compatForestry && Loader.isModLoaded(GTValues.FORESTRY)) {
 			GTMod.logger.info("Doing Forestry Things");
+			GTTileCentrifuge.addRecipe("gemApatite", 5, 0, GTTileCentrifuge.totalEu(4000), GTMaterialGen.getDust(GTMaterial.Calcite, 4), GTMaterialGen.getIc2(Ic2Items.fertilizer, 1));
 			GTRecipeForestry.notTheBees();
 		}
 	}